Rework diagnostic sums to be more customizable (#1452)
The current diagnostic sums functionality (castro.sum_interval = 1) has two limitations. First, customizing it for a problem setup currently means that the whole sum_integrated_quantities.cpp file has to be copied, which means that if the functionality from the main sum_integrated_quantities routine is desired, the problem has to duplicate that code. Second, even if you did customize it, the Amr class ownership of the data logs means that customizing the number and name of the logs is brittle. One unfortunate way that brittleness plays out is that you must specify amr.data_log in the inputs file with the exact ordering of the diagnostic names.

The intent of this change is that problems should have not have to override sum_integrated_quantities.cpp to get custom diagnostics. A new hook is provided, problem_diagnostics.H, which is called at the end of sum_integrated_quantities() for problem-specific diagnostics. To support this change, we stop depending on the Amr ownership of the data logs. Two new Castro class members are added, data_logs and problem_data_logs. The former is managed by Castro and can be changed over time if we want to add more diagnostic routines, with initialization done in Castro::read_params(). The latter is managed by the problem itself, with initialization done in problem_initialize(). The log names are coded directly into the source, which means that you don't need to remember to put those names in your inputs file.

test_convect and toy_convect are converted to the new scheme.