Move wdmerger initial model constructor to model_parser (#1454)
The isothermal initial model generator is moved from wdmerger into the C++ model parser. This will allow it to be reused multiple times in the coming split of wdmerger into separate problem setups for mergers, collisions, and TDEs. At some point we can also merge this with the initial model generator for the setups like flame_wave.