update the docs and scripts for bubble_convergence (#1478)
The bubble_convergence problem tests the 4th order SDC convergence with hydro, reactions, gravity, and reflecting BCs. It was shown in the Castro SDC paper. This PR updates the documentation to explain how to run it locally. It also fixes the main convergence script -- it was measuring the convergence at the midpoint in time and not at the end of the simulation, so it didn't agree with what was in the paper. With these updates, we get the same convergence that was printed in the paper. Finally, we add a 32^2 problem to explore the convergence at even lower resolution.