Convert SDC to C++ (#1208)
Converts all of the "true" SDC code paths to C++.  This integrates with the C++ VODE solver in Microphysics (if you are using VODE for the SDC solver).  For most of the test problems the agreement is O(1.e-8) or better.  bubble_convergence and reacting_convergence converge as 4th order, as expected.  You now need to specify that you want to use the numerical Jacobian for the VODE solver via integrator.jacobian=2 (it no longer uses castro.sdc_use_analytic_jac -- that only applies to the Newton solver).  The 4th order detonation problem is very sensitivity (and it is probably not a good idea to do a detonation 4th order because it is so discontinuous), and it works best with the numerical Jacobian + VODE.

Co-authored-by: Michael Zingale <michael.zingale@stonybrook.edu>
Co-authored-by: Max Katz <maxpkatz@gmail.com>