add axisymmetric geometric sources (#923)
In axisymmetric geometry, there are additional forces that arise due to the changing direction of the unit vectors in the div{rho U U} term. The paper by Bernand-Champmartin discusses this. See issue #913. Note that the way we handle the Coriolis force for 2-d axisymmetry is already consistent with a righthanded coordinate system even though we order the state as (r, z, theta).

Note: at the moment, this is not compatible with the idea from Byerly et al. to do rotation with the state data in the inertial frame, but that is off by default.