we don't seem to use the outflow data in Fortran anymore (#1669)
this looks like it was originally used for the case where we wanted to average the data on the grid into the boundaries to account for the case where we might not have an entire star in the domain.  But we don't actually use this.  This was the "spherical_star" functionality.