Use density-based sponge in DustCollapse (#1714)
DustCollapse uses a radius-based sponge, but since the radius of the spherical shell moves inward over time, the problem needs to recalculate the sponge radius at every timestep. This is what necessitated the update_sponge_params functionality.

After DustCollapse was implemented we added a density-based sponge. This more neatly solves the problem without requiring any problem-specific functionality. Switching to the density sponge for this problem will allow us to remove that custom sponge hook, which was only ever used for this problem.