Use S_new instead of S_old in CFL violation check (#1842)
The CFL violation check just prior to the hydro is intended to guess whether the hydro update will violate the CFL stability criterion. For that we can either use the old state or the state updated with the old-time sources. This change switches from the former to the latter. The motivation is the observation that when we fill new fine grids with data from coarse grids, the momentum may be inaccurate at a sharp density gradient like at the edge of a star. This change allows us to apply the sponge to the state (if a sponge is used) before checking whether it violates the CFL criterion, which will avoid spurious CFL violation warnings.