Protect against negative (rho e) in trans (#1850)
When using transverse_reset_rhoe, the code detects if the (rho E)-based transverse update makes (rho e) negative, and if so resets to an update that is based on the equation for (rho e). But it's possible for the (rho e) update to also result in a negative (rho e), particularly for ambient material. So we check against this case now and turn off the transverse correction if it would result in a negative (rho e). Also, if there was a density reset then we make (rho e) the unmodified edge state, similar to what we do with pressure.