Only warn on density reset if below retry_small_density_cutoff (#1848)
If the user sets castro.retry_small_density_cutoff > 0, they are telling the code not to care about density resets below this threshold. So the user probably does not want to be warned about these resets either; this change turns the warning off for that case. It can still be opted into with castro.v = 2.