Calculate position on grid for SCF relative to center (#1871)
This is a minor bug fix for the SCF calculation -- the positions on the grid should be calculated with respect to center because that's where the stellar edges are placed with respect to. This had no effect on the SCF single star test because the center for that problem happens to be at the origin.