Remove CFL violation check for CTU advance (#1873)
The code in question checks the old state (plus, as of #1842, the old-time sources) to see whether it violates the CFL criterion. But this has been superfluous ever since we added the logic at the end of the timestep to see if the timestep constraint is violated at that point. If there is a violation at that point, we will retry the timestep. So this prior additional check is now removed.