Initialize SDC reactions data to zero when do_react = 0 (#1874)
If we have compiled with reactions and USE_SIMPLIFIED_SDC, but have set castro.do_react = 0, there is a bug in the current code where the new-time Reactions_Type data doesn't actually get initialized to zero during the CTU advance. This means that if anything attempts to access the data, such as a derived variable, there may be nonsense results. This change sets the new-time reactions type to zero explicitly so that there will always be valid data to access. For safety, we also set the old-time Reactions_Type data, as well as the Simplified_SDC_React_Type data, to zero.