use the new initial model from initial_models/rotating_star (#1923)
This uses the new initial model from initial_models/massive_star. The main difference is that we reduce the MESA model down to 19 isotopes before putting into HSE and also preserve Ye from the MESA model and use that in the EOS in NSE regions.

The problem setup is also modified now to use Ye in regions where we are in NSE and use the X's otherwise, and it consistently initializes the other composition variables.