Zero out flux corrections that would cause small density (#1879)
As observed in #341, it's possible for the reflux operation to result in a small or negative density. Currently this is dealt with via a density reset. Density resets are fairly ugly and should be avoided since they can cause a large change in density (down to small_dens) that has downstream consequences. So instead what we can do is just zero out the flux corrections if they would cause a small or negative density. This PR applies that change. Another option would have been to use the existing limit_hydro_fluxes_on_small_dens() function, however this function's interface currently relies on having the primitive variables, which are not already available in the reflux operation. A future change could do this if desired.

This PR also modifies the update on the fluxes so that they are done on a direction-by-direction basis. While this is not strictly necessary to support this change, it should save on memory since we don't actually need to have all three directional fluxes constructed simultaneously.