Fix _prob_params parsing if a final newline is missing
If a line doesn't have a comment, then `line.find("#")` returns -1.
Normally, `line[:-1]` will just remove a trailing newline, but if the
file doesn't have a newline at the end, this will instead remove the
last character on the line, which could be a "y" in the 4th column for
letting a parameter be set at runtime (see #1980).