don't trace any SDC sources (#1965)
previously, the pressure and energy source terms were included in the PPM
tracing. Now we no longer do this and instead explicitly add them at the
end of all of the transverse updates. This has the advantage of letting
us check whether the interface states are negative and also removes
their influence from the reference states and eigensystem

Note: this also restores the pure hydro (non-SDC) species sources to the tracing
routines (they were taken out in #1646)