Prefetch radial gravity data before and after MPI reduction (#2091)
When GPU-aware MPI is not being used, it is not generally safe to pass managed memory directly to an MPI call (depending on MPI implementation, this could lead to accessing invalid data). So we need to prefetch the data to the host before calling ParallelDescriptor reduction operations. (This is handled already for multipole gravity, but this was missed when monopole gravity was ported to GPUs.)

This addresses the issue identified in #2080.