Fix box bounds in update_dcf (#2114)
Tiling was originally added to update_dcf in commit c54638f. However, I think there was a bug in this commit. The dcf array has a single ghost zone, but not all of the source arrays filling it do, and the original Fortran looped from dcf_lo+1 to dcf_hi-1, which would have been only the valid domain. But this commit used a growntilebox() for the MFIter loop, which is a change from the Fortran and one that will not make sense unless all of the source MultiFabs have a ghost zone too.