Radsolve default values should be set before overrides (#2113)
As originally written, this change unintentionally overrode the level_solver_flag in 1D, leading to diffs for inputs files that were not using Hypre SMG (radsolve.level_solver_flag = 0). To simplify the code and avoid the regression test diff, this override has been removed; there is already an abort later in this routine that covers the unsupported case radsolve.level_solver_flag = 1.