Simplify the density flux limiter algorithm (#2134)
The flux limiter that prevents small densities uses an algorithm that blends the second order flux with a first order flux which guarantees positivity of density (under certain conditions). However, this blending causes some undesirable behavior when it comes to multi-fluid advection, for example with the species fluxes potentially not moving in the same direction as the density flux which they are supposed to be passively following. So this change simplifies the algorithm to a simple linear scaling of the fluxes (which we had been doing anyway, and which I suspect this algorithm was simplifying to much of the time).