Add parameter for controlling GPU memory footprint (#2153)
The CTU hydro allocates a lot of temporary Fab data for its computation. This can result in a memory footprint on GPU builds that is significantly higher than the memory of the StateData itself. A new option, castro.hydro_memory_footprint_ratio, is added which can help limit the amount of temporary Fab data active at any one time.