Add new advecting white dwarf problem setup (#2168)
In this problem setup, a white dwarf is placed on a grid at a uniform translation velocity along the x-axis. The purpose of the problem is to study how well HSE is maintained in the advecting star, and also to understand the numerical issues at the edges of the star when it is moving at non-trivial speed relative to the domain. This is similar to the problem posed in Section 4.3.2 of Katz et al. (2016), which at the time was implemented as a variant of the wdmerger problem setup.