Fix how multiple faces are accounted for in density flux limiter (#2186)
As noted in #2182, the way we account for multiple faces is incorrect. We are currently increasing the floor by a factor of 6 (or 4 or 2, in 2D and 1D respectively). But instead we need to check on whether the flux multiplied by a factor of 6 would violate the floor. This is more aggressive and may result in some unnecessary limiting; in a future PR we can potentially be more sophisticated by considering the relative magnitude of the fluxes on the various faces.