Fix test for pre-advance timestep check (#2225)
This should run on everything but the very first fine timestep in a given coarse step.

Also, #2222 missed that estTimeStep always operated on the new-time state, so all of the estdt interfaces are updated to accept an is_new parameter and use the old-time state if it is set to zero. The pre-advance check then uses the old-time state as intended.