Only access old data when needed in GetCrsePhi (#2253)
On a restart, if castro.dump_old is not set to 1, there will not be old time data saved, including for PhiGrav_Type. But the code was attempting to access the old data on the multilevel Poisson solve after a restart in the linear interpolation in GetCrsePhi. This is unnecessary because in this solve after a restart, all levels are at the same time, so we can skip the interpolation and copy directly from the new-time state (using the same logic found in make_radial_gravity). So this works around the issue.