update flame_wave He inputs (#2267)
in particular, the 2-d version should use 1/2 the x size, since
it is axisymmetric.  Also fix the BCs for the pslope case in 3-d

this also adds an inputs with non-square zones
I did a 2-d XRB with this and it looked really good.
This is cooler than the hottest model we've used before, which
helps keep the flame structure clean.