remove derived in_nse plot var and rename new one to in_nse (#2409)
We switched from deriving whether we are in NSE for the plotfiles and instead
using the NSE state stored in the Reactions_Type StateData. This change eliminates
the duplicated derived in_nse plot variable and renames the new Reactions_Type
plot variable to in_nse. This makes our scripts compatible with the way of computing
it without needing to change them.