Re-implement the radial damping force that causes a wdmerger inspiral (#2471)
This implements the inspiral driving force from Pakmor et al. 2021. A force is applied in the direction of the stellar orbit that is intended to cause the orbital distance a to decrease at a uniform rate. The rate is determined as an arbitrary factor multiplied by the speed of the orbit (of the reduced mass version of the system). The inspiral force is shut off when the Roche radius of the secondary becomes smaller than some arbitrary multiple of the secondary's initial radius, as a way to approximate Roche lobe overflow.