On burn failure, return immediately before clean_state (#2521)
clean_state will abort in normalize_species() if the burn failed and left things in an unphysical situation, but we don't want to abort in this circumstance because the next step is to reject the advance and try again. So we avoid this by skipping the call to clean_state if we're going to be rejecting the advance.