include pressure in x-momentum flux in 1-d Cartesian (#2468)
previously, we were explicitly adding the pressure gradient in the conservative update
to the momentum for all 1-d geometries. But we were only doing the flux correction
on P_radial for non-Cartesian geometries. Now we treat all Cartesian geometries the
same, regardless of dimensionality.

This will likely change 1-d Cartesian answers.