fix outflow BCs for the 4th order solver (#2607)
In the interpolation of the interface states for the 4th order solver, we use a one-sided
stencil at the domain boundaries (so long as we are not periodic). This works great for
symmetry, and we further enforce the symmetry by reflecting the interface states.

For outflow BCs, we were trying to be fancy and continue to use this one-sided stencil
but extrapolate into the ghost cell to get a better prediction for the zone just outside of
the boundary. This does not work well and causes a flame problem to do bad things.

This PR removes all the fancy outflow attempts and has us only do a one-sided stencil
for reflecting. I think this is more correct, and will let us deal with inflow eventually,
although we will not be 4th order for outflow. But I am not sure an outflow BC is defined
well enough to understand what it means to be high order there.