clean up linear interpolation (#2664)
The locate() function was returning the index of the zone that is higher than the input coordinate r. The logic in interpolate accounted for this, but things were more complex than they need be. This simplifies locate and the linear interpolation.

It also adds a unit test to help understand what's happening.