Update documentation regarding species sources and normalization (#2742)
This PR updates the Hydrodynamics section documentation to highlight that sources for passive variables in the conserved state are off by default. It also removes the reference to castro.normalize_species, which does not appear to be a valid runtime parameter any longer. I am happy to introduce additional documentation regarding the species normalization if that would be helpful -- my understanding is that it is on by default now.