Compute the shock flag once -- at the start of the timestep (#2728)
This changes the way the shock flag is computed for disabling burning in shocks. Now we compute it
at the top of the timestep, before the first Strang burn, using the time-level n state and sources.
There is some redundancy here (we need to compute the old sources before the burn now),
and there is a FillPatch needed. But this should be minimal and only in effect if we compile with SHOCK_VAR.

Note: this does not change how we compute the shock flag for hydro at the moment
when doing the hybrid Riemann solve (although we can easily change that in a future PR).