Create problem to study He flame movement in a double det (#2870)
Begin creating a problem to study the lateral movement of a He flame in the accreted layer of a white dwarf.

We want to determine how various shock detection thresholds and burning treatments affect the initial flame movement in a double detonation model. In particular, we seek to see if the results of disabling/enabling burning in detected shock zones converges as resolution becomes finer.