only compute hydro version of shock variable for hybrid Riemann (#2865)
in the CTU solver, we compute a local version of the shock flag
used just for the Riemann solver.  This removes the SHOCK_VAR
check, since we only need this local version if we are doing the
hybrid Riemann solve.  This local shk array is never used for the
StateData USHK flag.