add geometric source terms to the PPM tracing (#2473)
Add the geometric source terms to the primitive variable sources so they can be
included in the tracing.

We can't simply add them in src_to_prim, since the sources depend on the
direction of the reconstruction. We also can't just add them in the trace_ppm
to srcQ immediately at the start and then do the reconstruction since we need
the geometric source terms in the ghost cells for the reconstruction. Therefore
we add these geometric sources to the 5-point stencil that will be traced just
before the tracing. This requires some redundant calculation, but it is not too
bad.