Only print mid-timestep estTimeStep outputs if relevant (#3023)
Output from estTimeStep is occurring for the pre- and post-timestep validity checks, but on most timesteps this clutters the output because the validity check passes. This PR creates a class that temporarily redirects std::cout and then uses that to capture the estTimeStep output and only print it if needed.