when we "make" without a "clean", ensure that the same net is used (#3073)
a common error is to build the executable, do a run, and then change
the network by building with NETWORK_DIR set, without a "make clean".
This leads to an undefined behavior.  This adds a check to ensure that
when we make we are using the same network as done originally, and
if not, aborts with a message saying to do "make clean"