Timestep estimator interfaces accept MultiFabs (#3122)
The estdt functions now take the MultiFab state they're working with rather than the is_new parameter. This enables a potential future change where we could estimate the timestep on some states that are different from the old and new time data. While doing this change, we also move the declarations to a separate header and update the namespace accordingly.