Combine estdt functions into a single template function (#3125)
The hydro, MHD, diffusion, and reactions timestep estimators are now combined into a single template function that uses an enumerator to select which physics we want to evaluate. Radiation is deferred to a later change since it is implemented differently.

This is setting up a future change where we can use a constexpr for loop to succinctly evaluate all of the timestep limiters in just a few lines of code. That will then enable a subsequent change where we can call the timestep limiter with multiple states inside the broader estTimeStep function; for example, we could do a first-order prediction to time-level n+1 and evaluate that state and then select our timestep as the minimum of those two values.